Remove leading zeros (#8)
The aim of this PR is to remove leading zeros after the country code.
We've seen a lot of issues where users don't understand that when placed
with a UI with a country code already there, they don't require any
leading zeros.

As an example:

 ```
 +4407481092544
 ```

The system at hand couldn't decipher that the user had inserted a zero
and rejected the number accordingly because of the range (length) check.

This fix removes all leading zeros after the country code, if there is
no country provided then it will not attempt to do this as the chances
are it would be wrong most of the time.

The removal of leading zeros is only on when turned on by use of a flag
on the PhoneData struct `RemoveLeadingZeros`. We've not turned this on
by default as there is a possibility that chaos could ensue, so it's
better to turn them on as we go!

* Invert the original default RemoveLeadingZeros

Instead of defaulting to `false`, we're going to flip and see what the
damage is.

When splitting the code, keep the left hand side rather than trying to
santize it! This way we keep the newly updated value as well